Create buttons.js fot the chuld component.
